Refactor: 예약이 존재하는 시간표 삭제 시도 시 서비스 레이어에서 명시적 예외 검증 처리

+++ src/main/java/roomescape/service/ReservationService.java
@@ -76,6 +76,9 @@ public class ReservationService {
         if (!timeRepository.existsById(id)) {
             throw new BadRequestException("삭제할 시간표 카운트가 부재합니다.");
         }
+        if (reservationRepository.existsByTimeId(id)) {
+            throw new BadRequestException("이미 해당 시간으로 예약된 내역이 존재하여 시간표를 삭제할 수 없습니다.");
+        }
         timeRepository.deleteById(id);
     }
 }
